+++ intelmq/bots/collectors/stomp/collector.py
@@ -8,7 +8,7 @@ try:
 except ImportError:
     stomp = None
 else:
-    class StompListener(stomp.listener.PrintingListener):
+    class StompListener(stomp.PrintingListener):
         """
         the stomp listener gets called asynchronously for
         every STOMP message
@@ -17,6 +17,7 @@ else:
             self.stompbot = n6stompcollector
             self.conn = conn
             self.destination = destination
+            super().__init__()
 
         def on_heartbeat_timeout(self):
             self.stompbot.logger.info("Heartbeat timeout. Attempting to re-connect.")
@@ -40,24 +41,27 @@ else:
             connect_and_subscribe(self.conn, self.stompbot.logger, self.destination)
 
 
-def connect_and_subscribe(conn, logger, destination):
-    conn.start()
-    connect_status = conn.connect(wait=True)
-    subscribe_status = conn.subscribe(destination=destination,
-                                      id=1, ack='auto')
-    logger.info('Successfully connected and subscribed. '
-                'Connect status: %r, subscribe status: %r.',
-                connect_status, subscribe_status)
+def connect_and_subscribe(conn, logger, destination, start=False):
+    if start:
+        conn.start()
+    conn.connect(wait=True)
+    conn.subscribe(destination=destination,
+                   id=1, ack='auto')
+    logger.info('Successfully connected and subscribed.')
 
 
 class StompCollectorBot(CollectorBot):
     """ main class for the STOMP protocol collector """
 
     collector_empty_process = True
+    conn = False  # define here so shutdown method can check for it
 
     def init(self):
         if stomp is None:
             raise ValueError('Could not import stomp. Please install it.')
+        elif stomp.__version__ < (4, 1, 8):
+            raise ValueError('stomp version >= 4.1.8 is required, please upgrade (installed: %s).'
+                             '' % '.'.join(map(str, stomp.__version__)))
 
         self.server = getattr(self.parameters, 'server', 'n6stream.cert.pl')
         self.port = getattr(self.parameters, 'port', 61614)
@@ -83,15 +87,16 @@ class StompCollectorBot(CollectorBot):
                                      ssl_key_file=self.ssl_cl_cert_key,
                                      ssl_cert_file=self.ssl_cl_cert,
                                      ssl_ca_certs=self.ssl_ca_cert,
-                                     wait_on_receipt=True,
                                      heartbeats=(self.heartbeat,
                                                  self.heartbeat))
 
         self.conn.set_listener('', StompListener(self, self.conn, self.exchange))
-        self.conn.start()
-        connect_and_subscribe(self.conn, self.logger, self.exchange)
+        connect_and_subscribe(self.conn, self.logger, self.exchange,
+                               start=stomp.__version__ < (4, 1, 20))
 
     def shutdown(self):
+        if not stomp or not self.conn:
+            return
         try:
             self.conn.disconnect()
         except stomp.exception.NotConnectedException: