+++ CHANGELOG.md
@@ -20,6 +20,7 @@ CHANGELOG
 ### Bots
 #### Collectors
 - `intelmq.bots.collectors.mail.collector_mail_attach`: fix handling of deprecated parameter name `attach_unzip`.
+- `intelmq.bots.collectors.stomp.collector`: Fix compatibility with stomp.py versions `> 4.1.20` and catch errors on shutdown.
 
 #### Parsers
 - `intelmq.bots.parser.cymru.parser_cap_program`: Support for protocol 11 (`nvp-ii`).